extract a couple types
