+++ lib/maplibre-compose/src/commonMain/kotlin/dev/sargunv/maplibrecompose/core/expression/Expression.kt
@@ -1,10 +1,5 @@
 package dev.sargunv.maplibrecompose.core.expression
 
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import dev.sargunv.maplibrecompose.core.expression.Expressions.const
-
 /** Wraps a JSON-like value that represents an expression, typically used for styling map layers. */
 public sealed interface Expression {
   /** The JSON-like value that backs this expression. */
@@ -105,99 +100,3 @@ public sealed interface ComparableExpression : Expression
  * [ExpressionScope.interpolate].
  */
 public sealed interface InterpolateableExpression : Expression
-
-public sealed interface Interpolation : Expression {
-
-  /** Interpolates linearly between the pairs of stops. */
-  public data object Linear : Interpolation by ExpressionImpl.ofString("linear")
-
-  /**
-   * Interpolates exponentially between the stops.
-   *
-   * @param [base] controls the rate at which the output increases: higher values make the output
-   *   increase more towards the high end of the range. With values close to 1 the output increases
-   *   linearly.
-   */
-  public class Exponential(base: FloatExpression) :
-    Interpolation by ExpressionImpl.ofList(listOf(const("exponential"), base))
-
-  /**
-   * Interpolates using the cubic bezier curve defined by the given control points between the pairs
-   * of stops.
-   */
-  public class CubicBezier(
-    x1: FloatExpression,
-    y1: FloatExpression,
-    x2: FloatExpression,
-    y2: FloatExpression,
-  ) : Interpolation by ExpressionImpl.ofList(listOf(const("cubic-bezier"), x1, y1, x2, y2))
-}
-
-internal class ExpressionImpl private constructor(override val value: Any?) :
-  BooleanExpression,
-  FloatExpression,
-  IntExpression,
-  DpExpression,
-  StringExpression,
-  EnumExpression<LayerPropertyEnum>,
-  ColorExpression,
-  MapExpression,
-  VectorExpression, // also provides ListExpression<*> (with type erasure)
-  OffsetExpression,
-  DpOffsetExpression,
-  PaddingExpression,
-  FormattedExpression,
-  ImageExpression,
-  CollatorExpression,
-  Interpolation {
-  internal inline fun <reified T : Expression> cast(): T = this as T
-
-  internal companion object {
-    val ofNull = ExpressionImpl(null)
-    val ofTrue = ExpressionImpl(true)
-    val ofFalse = ExpressionImpl(false)
-
-    private val constSmallInts = Array(512) { ExpressionImpl(it.toFloat()) }
-    private val constSmallFloats = Array(512) { ExpressionImpl(it.toFloat() / 20f) }
-    private val ofEmptyString = ExpressionImpl("")
-    private val ofTransparent = ExpressionImpl(Color.Transparent)
-    private val ofBlack = ExpressionImpl(Color.Black)
-    private val ofWhite = ExpressionImpl(Color.White)
-    private val ofEmptyList = ExpressionImpl(emptyList<Any?>())
-    private val ofZeroOffset = ExpressionImpl(Offset.Zero)
-    private val ofZeroPadding = ExpressionImpl(ZeroPadding)
-
-    private val ofEmptyMap = ExpressionImpl(emptyMap<String, Any?>())
-
-    private fun isSmallInt(f: Float) = f >= 0 && f < 512 && f.toInt().toFloat() == f
-
-    fun ofFloat(float: Float) =
-      when {
-        isSmallInt(float) -> constSmallInts[float.toInt()]
-        isSmallInt(float * 20f) -> constSmallFloats[(float * 20f).toInt()]
-        else -> ExpressionImpl(float)
-      }
-
-    fun ofString(string: String) = if (string.isEmpty()) ofEmptyString else ExpressionImpl(string)
-
-    fun ofColor(color: Color) =
-      when (color) {
-        Color.Transparent -> ofTransparent
-        Color.Black -> ofBlack
-        Color.White -> ofWhite
-        else -> ExpressionImpl(color)
-      }
-
-    fun ofMap(map: Map<String, Expression>) =
-      if (map.isEmpty()) ofEmptyMap else ExpressionImpl(map.mapValues { it.value.value })
-
-    fun ofList(list: List<Expression>) =
-      if (list.isEmpty()) ofEmptyList else ExpressionImpl(list.map { it.value })
-
-    fun ofOffset(offset: Offset) =
-      if (offset == Offset.Zero) ofZeroOffset else ExpressionImpl(offset)
-
-    fun ofPadding(padding: PaddingValues.Absolute) =
-      if (padding == ZeroPadding) ofZeroPadding else ExpressionImpl(padding)
-  }
-}
